picolibc: Avoid build system headers
When building picolibc as a companion library the configure step can end
up picking up certain headers from the build system which causes build
failures. Pass an appropriate --with-headers= to the GCC back end when
building picolibc as a companion library so that the correct headers are
detected by ./configure.

Signed-off-by: Chris Packham <judge.packham@gmail.com>