m4/ctng_python.m4: add python 3.13 to searched versions
When trying to build and run ct-ng locally on Archlinux, we get plenty
of warnings interleaved with the expected output, right at the
list-samples step:

  [...]
  config/configure.in:54:warning: config symbol defined without type
  [L...]   aarch64-ol7u9-linux-gnu
  config/configure.in:54:warning: config symbol defined without type
  [L...]   aarch64-ol8u6-linux-gnu
  config/configure.in:54:warning: config symbol defined without type
  [L...]   aarch64-ol8u7-linux-gnu
  config/configure.in:54:warning: config symbol defined without type
  [L...]   aarch64-rpi3-linux-gnu
  config/configure.in:54:warning: config symbol defined without type
  [...]

The corresponding config/configure.in file indeed shows that the KConfig
variable for python >= 3.4 is indeed missing some default value:

  [...]
  config CONFIGURE_has_gnu_m4_1_4_12_or_newer
      def_bool y

  config CONFIGURE_has_python_3_4_or_newer

  config CONFIGURE_has_bison_2_7_or_newer
      def_bool y
  [...]

The m4 script in charge of detecting a valid python version, having an
explicit list of supported python version, does not have the latest
python version (3.13, released in October, 2024), which happens to be
the current default python version on Archlinux.

Fix ct-ng list-samples output by adding this 3.13 version in the list of
tested versions.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>