CI: Use wget --no-use-server-timestamps
When downloading files wget will set the timestamp on the local file to
match the timestamp on the server. For the purpose of maintaining the
cached files for the CI build we need the timestamp to be updated if the
file is downloaded. Use the --no-use-server-timestamps option for wget
so that the local file timestamp reflects the download time.

Signed-off-by: Chris Packham <judge.packham@gmail.com>