companion_libs: Apply EXTRA_CFLAGS to target build only
Commit 93d5d766 ("Use -std=c17 for companion libs, which won't build
with C23 standard, default for gcc15") was attempting to avoid having to
play whack-a-mole with gmp and ncurses due to GCC 15 defaulting to
-std=gnu23. Unfortunately the extra CFLAGS were applied generically so
also affected the build & host compilation of the packages. This caused
problems for older build machines which had versions of GCC that do not
understand -std=gnu17.

As the kconfig check GCC_15_or_later only applies to the target anyway
move the use of EXTRA_CFLAGS to the target build only.

It is likely that we might need to do something equivalent for the build
& host compiles as more distros pick up GCC 15 but that would need to be
accompanied by a check of the build/host compiler.

Fixes #2463
Signed-off-by: Chris Packham <judge.packham@gmail.com>